# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Uni Shop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  university: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  university: string;
  bio?: string;
  avatarUrl?: string;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  available: 'available',
  sold: 'sold',
  reserved: 'reserved',
} as const;

export interface Listing {
  id: number;
  title: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  sellerId: number;
  sellerName: string;
  sellerUniversity: string;
  status: ListingStatus;
  /** @nullable */
  tags?: string | null;
  favoriteCount: number;
  createdAt: string;
}

export interface ListingInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  price: number;
  category: string;
  imageUrl?: string;
  sellerId: number;
  tags?: string;
}

export type ListingUpdateStatus = typeof ListingUpdateStatus[keyof typeof ListingUpdateStatus];


export const ListingUpdateStatus = {
  available: 'available',
  sold: 'sold',
  reserved: 'reserved',
} as const;

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  status?: ListingUpdateStatus;
  tags?: string;
}

export interface FavoriteInput {
  userId: number;
  listingId: number;
}

export interface FavoriteResult {
  favorited: boolean;
}

export interface ContactInput {
  name: string;
  email: string;
  message: string;
}

export interface ContactResult {
  success: boolean;
}

export interface MarketplaceStats {
  totalListings: number;
  totalUsers: number;
  totalCategories: number;
  totalSold: number;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name?: string;
  size?: number;
  contentType?: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: RequestUploadUrlResponseMetadata;
}

export type GetListingsParams = {
category?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
sort?: GetListingsSort;
};

export type GetListingsSort = typeof GetListingsSort[keyof typeof GetListingsSort];


export const GetListingsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  popular: 'popular',
} as const;

export type GetFavoritesParams = {
userId: number;
};