# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Uni Shop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all listings
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a single listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /listings/{id}/contact:
    post:
      operationId: contactSeller
      tags: [listings]
      summary: Send a message to the seller of a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Contact request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get all favorite listings for a user (by userId query param)
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorite listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /favorites/toggle:
    post:
      operationId: toggleFavorite
      tags: [favorites]
      summary: Toggle a listing as favorite for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "200":
          description: Toggle result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResult"
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{id}/listings:
    get:
      operationId: getUserListings
      tags: [users]
      summary: Get all listings by a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, university, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        university:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, university]
      properties:
        name:
          type: string
          minLength: 1
        university:
          type: string
          minLength: 1
        bio:
          type: string
        avatarUrl:
          type: string
    Listing:
      type: object
      required: [id, title, description, price, category, sellerId, sellerName, sellerUniversity, status, createdAt, favoriteCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerUniversity:
          type: string
        status:
          type: string
          enum: [available, sold, reserved]
        tags:
          type: ["string", "null"]
        favoriteCount:
          type: integer
        createdAt:
          type: string
    ListingInput:
      type: object
      required: [title, description, price, category, sellerId]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        sellerId:
          type: integer
        tags:
          type: string
    ListingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          enum: [available, sold, reserved]
        tags:
          type: string
    FavoriteInput:
      type: object
      required: [userId, listingId]
      properties:
        userId:
          type: integer
        listingId:
          type: integer
    FavoriteResult:
      type: object
      required: [favorited]
      properties:
        favorited:
          type: boolean
    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string
    ContactResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...